@@ -1,12 +1,13 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
+import PropTypes from 'prop-types';
 import InfiniteGrid from '../src/grid';
 
 class ExampleItem extends React.Component {
 
   static get propTypes() {
     return {
-      index: React.PropTypes.number
+      index: PropTypes.number
     };
   }
 
