@@ -1,4 +1,5 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 import {isEqual} from 'lodash';
 import Item from './item';
 
@@ -6,15 +7,15 @@ export default class InfiniteGrid extends React.Component {
 
 	static get propTypes() {
 		return {
-			itemClassName: React.PropTypes.string,
-			entries: React.PropTypes.arrayOf(React.PropTypes.object).isRequired,
-			height: React.PropTypes.number,
-			width: React.PropTypes.number,
-			padding: React.PropTypes.number,
-			wrapperHeight: React.PropTypes.number,
-			lazyCallback: React.PropTypes.func,
-			renderRangeCallback: React.PropTypes.func,
-			buffer: React.PropTypes.number
+			itemClassName: PropTypes.string,
+			entries: PropTypes.arrayOf(PropTypes.object).isRequired,
+			height: PropTypes.number,
+			width: PropTypes.number,
+			padding: PropTypes.number,
+			wrapperHeight: PropTypes.number,
+			lazyCallback: PropTypes.func,
+			renderRangeCallback: PropTypes.func,
+			buffer: PropTypes.number
 		}
 	}
 
